Refactor Lighting and Swag components; move files to new directories and update imports
