+++ src/App.js
@@ -13,13 +13,13 @@ import FAQ from "./pages/FAQ/FreqQues.jsx";
 import QuoteRequest from './pages/QuoteRequest/QuoteRequest';
 import AboutUs from './pages/AboutUs/AboutUs';
 import WhiteFrameTents from './pages/WhiteFrameTents/WhiteFrameTents';
-import ClearTopTents from './pages/ClearTopTents/ClearTopTents';
+import ClearTopTents from './pages/TentStyles/ClearTopTents/ClearTopTents.jsx';
 import TentLiner from './pages/TentLiner/TentLiner';
 import Skylight from './pages/Skylight/Skylight';
 import Sailcloth from './pages/Sailcloth/Sailcloth';
 import Swag from './pages/Swag/Swag';
 import TentPackages from './pages/TentPackages/TentPackages';
-import TentLinerPackages from './pages/TentLinerPackages/TentLinerPackages';
+import TentLinerPackages from './pages/Packages/TentLinerPackages/TentLinerPackages.jsx';
 import ClearTopPackages from './pages/ClearTopPackages/ClearTopPackages';
 import SailclothPackages from './pages/SailclothPackages/SailclothPackages';
 import RusticPackages from './pages/RusticPackages/RusticPackages';
@@ -31,7 +31,7 @@ import Flooring from './pages/Flooring/Flooring';
 import Bars from './pages/Bars/Bars';
 import BackdropsAndDraping from './pages/BackdropsAndDraping/BackdropsAndDraping';
 import AVStaging from './pages/AVStaging/AVStaging';
-import HeatingAndCooling from './pages/HeatingAndCooling/HeatingAndCooling';
+import HeatingAndCooling from './pages/Accessories/HeatingAndCooling/HeatingAndCooling.jsx';
 import Login from "./admin/login/Login.jsx";
 import CarRental from "./admin/CarRental/CarRental.jsx";
 // import  PartyRental from "./admin/partyRental/PartyRental.jsx";
