+++ src/pages/Accessories/HeatingAndCooling/HeatingAndCooling.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import Navbar from '../../components/PartyRentalsPage/Navbar';
-import Footer from '../../components/Footer/Footer';
+import Navbar from '../../../components/PartyRentalsPage/Navbar';
+import Footer from '../../../components/Footer/Footer';
 
 const HeatingAndCooling = () => {
   const heatingCoolingOptions = [
