+++ src/pages/Accessories/Lighting/Lighting.jsx
