+++ src/pages/Packages/TentLinerPackages/TentLinerPackages.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Phone, Mail, MapPin, Calendar, ArrowRight } from 'lucide-react';
-import Navbar from '../../components/PartyRentalsPage/Navbar';
-import Footer from '../../components/Footer/Footer';
+import Navbar from '../../../components/PartyRentalsPage/Navbar';
+import Footer from '../../../components/Footer/Footer';
 
 const TentLinerPackages = () => {
   return (
