+++ src/pages/TentStyles/ClearTopTents/ClearTopTents.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Tent } from 'lucide-react';
-import Navbar from '../../components/PartyRentalsPage/Navbar';
-import Footer from '../../components/Footer/Footer';
+import Navbar from '../../../components/PartyRentalsPage/Navbar';
+import Footer from '../../../components/Footer/Footer';
 
 const ClearTopTents = () => {
   const images = [
