+++ src/pages/TentStyles/Sailcloth/Sailcloth.jsx
