+++ src/pages/TentStyles/Skylight/Skylight.jsx
