+++ src/pages/TentStyles/WhiteFrameTents/WhiteFrameTents.jsx
